refactor(shakemaps): move shakemap code into folder and split files
